CoalNet/ml-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
statsmodels==0.14.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
